Fix up ct_value so it works as expected when the requested type if not the same as the exp type (ie, try to cast).
svn-id: zl/trunk@181

+++ expand.cpp
@@ -1422,5 +1422,5 @@ Flags::Flags(ChangeSrc<T> & f, const Flags & o) {
 
 unsigned long ct_value(const Syntax * p, Environ * env) {
   AST * ast = parse_exp(p, *env);
-  return ast->ct_value<unsigned int>();
+  return ast->ct_value<unsigned long>();
 }
